requirements.txt
altgraph==0.17.2
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
comtypes==1.1.10
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.51.0
django-ckeditor==6.4.2
django-etc==1.3.1
django-extensions==3.2.0
django-hitcount==1.3.5
django-js-asset==2.0.0
django-mptt==0.13.4
django-social-share==2.3.0
django-sslserver==0.22
future==0.18.2
idna==3.3
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
Werkzeug==2.2.0
